# api-spec/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/lesson/{id}":{"get":{"tags":["lesson-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lesson"}}}}}},"put":{"tags":["lesson-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonBase"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lesson"}}}}}},"delete":{"tags":["lesson-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/lesson":{"get":{"tags":["lesson-controller"],"operationId":"get","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Lesson"}}}}}}},"post":{"tags":["lesson-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLessonRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lesson"}}}}}}}},"components":{"schemas":{"LessonBase":{"required":["description","title"],"type":"object","properties":{"title":{"maxLength":255,"minLength":0,"type":"string"},"description":{"maxLength":255,"minLength":0,"type":"string"}}},"Lesson":{"required":["dateCreated","dateLastUpdated","description","title"],"type":"object","properties":{"title":{"maxLength":255,"minLength":0,"type":"string"},"description":{"maxLength":255,"minLength":0,"type":"string"},"id":{"type":"integer","format":"int64"},"dateCreated":{"type":"string","format":"date"},"dateLastUpdated":{"type":"string","format":"date"}}},"CreateLessonRequest":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}}}}}}